move file to the right location
